+++ tests/test_pytorch_helpers.py
@@ -1,6 +1,5 @@
 """Tests for `mlbench_core.utils.pytorch.helpers` package."""
 
-from mlbench_core.utils import LogMetrics
 from mlbench_core.utils.pytorch.helpers import config_path, config_pytorch
 
 
@@ -16,16 +15,6 @@ def test_config_pytorch(mocker):
     assert rank == 1
     assert world_size == 1
     assert graph is not None
-
-
-def test_LogMetrics(mocker):
-    mocker.patch("mlbench_core.api.ApiClient")
-
-    LogMetrics.log("1", 1, 1, "loss", 123)
-
-    mocker.patch.dict("os.environ", {"MLBENCH_IN_DOCKER": "True"})
-
-    LogMetrics.log("1", 1, 1, "loss", 123)
 
 
 def test_config_path(mocker):
